precios-luque/requirements.txt
aenum==3.1.15
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
dbf==0.99.9
Django==5.1.4
et_xmlfile==2.0.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.3.1
openpyxl==3.1.5
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.5.0
